:recycle: Orchestrator + refactoring
* Implement an orchestrator to build PurchaseRecap using CSV
* Call node API using object built by the CSV file
* Remove unused classes and methods

+++ backend-java/src/test/java/com/codeworks/onboarding/infrastructure/purchase/items/PurchaseItemsServiceImplTest.java
@@ -33,14 +33,14 @@ public class PurchaseItemsServiceImplTest {
         Assert.assertEquals(Long.valueOf(1L), purchaseItem.getBuyerId());
         Assert.assertEquals("pencil", purchaseItem.getLabel());
         Assert.assertEquals(Integer.valueOf(10), purchaseItem.getQuantity());
-        Assert.assertEquals(Integer.valueOf(20), purchaseItem.getUnitPrice());
+        Assert.assertEquals(Double.valueOf(20d), purchaseItem.getUnitPrice());
         Assert.assertEquals(4, purchaseItems.size());
     }
 
     @Test
     public void should_get_purchase_item_by_id() {
         when(purchaseItemsRepository.findById(Mockito.anyLong()))
-                .thenReturn(PurchaseItemsEntity.builder().id(1L).purchaseId(1L).buyerId(1L).label("pencil").quantity(10).unitPrice(20).build());
+                .thenReturn(PurchaseItemsEntity.builder().id(1L).purchaseId(1L).buyerId(1L).label("pencil").quantity(10).unitPrice(20d).build());
 
         PurchaseItemsEntity purchaseItem = purchaseItemsService.findPurchaseItemBy(1L);
 
@@ -49,21 +49,21 @@ public class PurchaseItemsServiceImplTest {
         Assert.assertEquals(Long.valueOf(1L), purchaseItem.getBuyerId());
         Assert.assertEquals("pencil", purchaseItem.getLabel());
         Assert.assertEquals(Integer.valueOf(10), purchaseItem.getQuantity());
-        Assert.assertEquals(Integer.valueOf(20), purchaseItem.getUnitPrice());
+        Assert.assertEquals(Double.valueOf(20d), purchaseItem.getUnitPrice());
     }
 
     @Test
     public void should_save_purchase_item() {
-        when(purchaseItemsRepository.save(Mockito.any())).thenReturn(PurchaseItemsEntity.builder().id(1L).purchaseId(1L).buyerId(1L).label("pencil").quantity(10).unitPrice(20).build());
+        when(purchaseItemsRepository.save(Mockito.any())).thenReturn(PurchaseItemsEntity.builder().id(1L).purchaseId(1L).buyerId(1L).label("pencil").quantity(10).unitPrice(20d).build());
 
-        PurchaseItemsEntity purchaseItem = purchaseItemsService.create(PurchaseItemsEntity.builder().id(1L).purchaseId(1L).buyerId(1L).label("pencil").quantity(10).unitPrice(20).build());
+        PurchaseItemsEntity purchaseItem = purchaseItemsService.create(PurchaseItemsEntity.builder().id(1L).purchaseId(1L).buyerId(1L).label("pencil").quantity(10).unitPrice(20d).build());
 
         Assert.assertEquals(Long.valueOf(1L), purchaseItem.getId());
         Assert.assertEquals(Long.valueOf(1L), purchaseItem.getPurchaseId());
         Assert.assertEquals(Long.valueOf(1L), purchaseItem.getBuyerId());
         Assert.assertEquals("pencil", purchaseItem.getLabel());
         Assert.assertEquals(Integer.valueOf(10), purchaseItem.getQuantity());
-        Assert.assertEquals(Integer.valueOf(20), purchaseItem.getUnitPrice());
+        Assert.assertEquals(Double.valueOf(20d), purchaseItem.getUnitPrice());
     }
 
     @Test
@@ -77,10 +77,10 @@ public class PurchaseItemsServiceImplTest {
 
     private List<PurchaseItemsEntity> buildPurchaseItems() {
         return Arrays.asList(
-                PurchaseItemsEntity.builder().id(1L).purchaseId(1L).buyerId(1L).label("pencil").quantity(10).unitPrice(20).build(),
-                PurchaseItemsEntity.builder().id(2L).purchaseId(1L).buyerId(1L).label("pencil").quantity(10).unitPrice(20).build(),
-                PurchaseItemsEntity.builder().id(3L).purchaseId(1L).buyerId(1L).label("pencil").quantity(10).unitPrice(20).build(),
-                PurchaseItemsEntity.builder().id(4L).purchaseId(1L).buyerId(1L).label("pencil").quantity(10).unitPrice(20).build()
+                PurchaseItemsEntity.builder().id(1L).purchaseId(1L).buyerId(1L).label("pencil").quantity(10).unitPrice(20d).build(),
+                PurchaseItemsEntity.builder().id(2L).purchaseId(1L).buyerId(1L).label("pencil").quantity(10).unitPrice(20d).build(),
+                PurchaseItemsEntity.builder().id(3L).purchaseId(1L).buyerId(1L).label("pencil").quantity(10).unitPrice(20d).build(),
+                PurchaseItemsEntity.builder().id(4L).purchaseId(1L).buyerId(1L).label("pencil").quantity(10).unitPrice(20d).build()
         );
     }
 }
